moved export_har to folder and provided tests for exporthar.py
