+++ examples/contrib/export_har/flows/correct_flows.har
@@ -1,0 +1,199 @@
+{
+    "log": {
+        "version": "1.2",
+        "creator": {
+            "name": "mitmproxy exporthar",
+            "version": "0.1",
+            "comment": "mitmproxy version mitmproxy 9.0.1"
+        },
+        "pages": [],
+        "entries": [
+            {
+                "startedDateTime": "2023-07-17T13:04:44.056651+00:00",
+                "time": 0,
+                "request": {
+                    "method": "GET",
+                    "url": "https://example.com/",
+                    "httpVersion": "HTTP/1.1",
+                    "cookies": [],
+                    "headers": [
+                        {
+                            "name": "content-length",
+                            "value": "0"
+                        },
+                        {
+                            "name": "Host",
+                            "value": "example.com"
+                        }
+                    ],
+                    "queryString": [],
+                    "headersSize": 61,
+                    "bodySize": 0
+                },
+                "response": {
+                    "status": 0,
+                    "statusText": "",
+                    "httpVersion": "",
+                    "headers": [],
+                    "cookies": [],
+                    "redirectURL": "",
+                    "headersSize": -1,
+                    "bodySize": -1,
+                    "_transferSize": 0,
+                    "_error": null,
+                    "content": []
+                },
+                "cache": {},
+                "timings": {
+                    "send": 0,
+                    "receive": 0,
+                    "wait": 0,
+                    "connect": 0,
+                    "ssl": -1
+                }
+            },
+            {
+                "startedDateTime": "2023-07-17T14:35:39.872231+00:00",
+                "time": 0,
+                "request": {
+                    "method": "GET",
+                    "url": "https://google.com/",
+                    "httpVersion": "HTTP/1.1",
+                    "cookies": [],
+                    "headers": [
+                        {
+                            "name": "content-length",
+                            "value": "0"
+                        },
+                        {
+                            "name": "Host",
+                            "value": "google.com"
+                        }
+                    ],
+                    "queryString": [],
+                    "headersSize": 60,
+                    "bodySize": 0
+                },
+                "response": {
+                    "status": 0,
+                    "statusText": "",
+                    "httpVersion": "",
+                    "headers": [],
+                    "cookies": [],
+                    "redirectURL": "",
+                    "headersSize": -1,
+                    "bodySize": -1,
+                    "_transferSize": 0,
+                    "_error": null,
+                    "content": []
+                },
+                "cache": {},
+                "timings": {
+                    "send": 0,
+                    "receive": 0,
+                    "wait": 0,
+                    "connect": 0,
+                    "ssl": -1
+                }
+            },
+            {
+                "startedDateTime": "2023-07-17T14:35:51.329621+00:00",
+                "time": 0,
+                "request": {
+                    "method": "POST",
+                    "url": "https://google.com/",
+                    "httpVersion": "HTTP/1.1",
+                    "cookies": [],
+                    "headers": [
+                        {
+                            "name": "content-length",
+                            "value": "0"
+                        },
+                        {
+                            "name": "Host",
+                            "value": "google.com"
+                        }
+                    ],
+                    "queryString": [],
+                    "headersSize": 60,
+                    "bodySize": 0,
+                    "postData": {
+                        "mimeType": "",
+                        "text": "",
+                        "params": []
+                    }
+                },
+                "response": {
+                    "status": 0,
+                    "statusText": "",
+                    "httpVersion": "",
+                    "headers": [],
+                    "cookies": [],
+                    "redirectURL": "",
+                    "headersSize": -1,
+                    "bodySize": -1,
+                    "_transferSize": 0,
+                    "_error": null,
+                    "content": []
+                },
+                "cache": {},
+                "timings": {
+                    "send": 0,
+                    "receive": 0,
+                    "wait": 0,
+                    "connect": 0,
+                    "ssl": -1
+                }
+            },
+            {
+                "startedDateTime": "2023-07-17T14:36:02.914265+00:00",
+                "time": 0,
+                "request": {
+                    "method": "POST",
+                    "url": "https://google.com/",
+                    "httpVersion": "HTTP/1.1",
+                    "cookies": [],
+                    "headers": [
+                        {
+                            "name": "content-length",
+                            "value": "0"
+                        },
+                        {
+                            "name": "Host",
+                            "value": "google.com"
+                        }
+                    ],
+                    "queryString": [],
+                    "headersSize": 60,
+                    "bodySize": 0,
+                    "postData": {
+                        "mimeType": "",
+                        "text": "",
+                        "params": []
+                    }
+                },
+                "response": {
+                    "status": 0,
+                    "statusText": "",
+                    "httpVersion": "",
+                    "headers": [],
+                    "cookies": [],
+                    "redirectURL": "",
+                    "headersSize": -1,
+                    "bodySize": -1,
+                    "_transferSize": 0,
+                    "_error": null,
+                    "content": []
+                },
+                "cache": {},
+                "timings": {
+                    "send": 0,
+                    "receive": 0,
+                    "wait": 0,
+                    "connect": 0,
+                    "ssl": -1
+                }
+            }
+        ]
+    }
+}
