+++ examples/contrib/export_har/flows/logfile
@@ -1,0 +1,1 @@
+1398:4:type;4:http;7:version;2:18#9:websocket;0:~8:response;0:~7:request;291:4:path;1:/,9:authority;0:,6:scheme;5:https,6:method;3:GET,4:port;3:443#4:host;11:example.com;13:timestamp_end;17:1689599084.056651^15:timestamp_start;17:1689599084.056651^8:trailers;0:~7:content;0:,7:headers;52:22:14:content-length,1:0,]22:4:Host,11:example.com,]]12:http_version;8:HTTP/1.1,}17:timestamp_created;17:1689599084.061592^7:comment;0:;8:metadata;0:}6:marked;0:;9:is_replay;0:~11:intercepted;5:false!11:server_conn;409:4:via2;0:~11:cipher_list;0:]11:cipher_name;0:~11:alpn_offers;0:]16:certificate_list;0:]3:tls;5:false!5:error;0:~5:state;1:0#3:via;0:~11:tls_version;0:~15:tls_established;5:false!19:timestamp_tls_setup;0:~19:timestamp_tcp_setup;0:~15:timestamp_start;0:~13:timestamp_end;0:~14:source_address;0:~3:sni;0:~10:ip_address;0:~2:id;36:5bcc6919-bec0-4201-b2a8-81d37aa04dc8;4:alpn;0:~7:address;21:11:example.com;3:443#]}11:client_conn;410:10:proxy_mode;7:regular;11:cipher_list;0:]11:alpn_offers;0:]16:certificate_list;0:]3:tls;5:false!5:error;0:~8:sockname;7:0:;1:0#]5:state;1:3#11:tls_version;0:~14:tls_extensions;0:]15:tls_established;5:false!19:timestamp_tls_setup;0:~15:timestamp_start;18:1689599084.0565512^13:timestamp_end;0:~3:sni;0:~8:mitmcert;0:~2:id;36:592af050-4844-4a8b-882c-b22fa2ddfc78;11:cipher_name;0:~4:alpn;0:~7:address;7:0:;1:0#]}5:error;0:~2:id;36:25e6bb05-b86a-49fb-b514-c5d515378744;}1394:4:type;4:http;7:version;2:18#9:websocket;0:~8:response;0:~7:request;289:4:path;1:/,9:authority;0:,6:scheme;5:https,6:method;3:GET,4:port;3:443#4:host;10:google.com;13:timestamp_end;17:1689604539.872232^15:timestamp_start;17:1689604539.872231^8:trailers;0:~7:content;0:,7:headers;51:22:14:content-length,1:0,]21:4:Host,10:google.com,]]12:http_version;8:HTTP/1.1,}17:timestamp_created;17:1689604539.873284^7:comment;0:;8:metadata;0:}6:marked;0:;9:is_replay;0:~11:intercepted;5:false!11:server_conn;408:4:via2;0:~11:cipher_list;0:]11:cipher_name;0:~11:alpn_offers;0:]16:certificate_list;0:]3:tls;5:false!5:error;0:~5:state;1:0#3:via;0:~11:tls_version;0:~15:tls_established;5:false!19:timestamp_tls_setup;0:~19:timestamp_tcp_setup;0:~15:timestamp_start;0:~13:timestamp_end;0:~14:source_address;0:~3:sni;0:~10:ip_address;0:~2:id;36:2390398b-e903-4cf2-aa39-37825231257d;4:alpn;0:~7:address;20:10:google.com;3:443#]}11:client_conn;409:10:proxy_mode;7:regular;11:cipher_list;0:]11:alpn_offers;0:]16:certificate_list;0:]3:tls;5:false!5:error;0:~8:sockname;7:0:;1:0#]5:state;1:3#11:tls_version;0:~14:tls_extensions;0:]15:tls_established;5:false!19:timestamp_tls_setup;0:~15:timestamp_start;17:1689604539.872131^13:timestamp_end;0:~3:sni;0:~8:mitmcert;0:~2:id;36:21ea5735-64d1-499d-9e93-a8edb451d807;11:cipher_name;0:~4:alpn;0:~7:address;7:0:;1:0#]}5:error;0:~2:id;36:6a0c192f-f4ef-481a-b35c-f0b78e323e28;}1396:4:type;4:http;7:version;2:18#9:websocket;0:~8:response;0:~7:request;290:4:path;1:/,9:authority;0:,6:scheme;5:https,6:method;4:POST,4:port;3:443#4:host;10:google.com;13:timestamp_end;17:1689604551.329621^15:timestamp_start;17:1689604551.329621^8:trailers;0:~7:content;0:,7:headers;51:22:14:content-length,1:0,]21:4:Host,10:google.com,]]12:http_version;8:HTTP/1.1,}17:timestamp_created;17:1689604551.330397^7:comment;0:;8:metadata;0:}6:marked;0:;9:is_replay;0:~11:intercepted;5:false!11:server_conn;408:4:via2;0:~11:cipher_list;0:]11:cipher_name;0:~11:alpn_offers;0:]16:certificate_list;0:]3:tls;5:false!5:error;0:~5:state;1:0#3:via;0:~11:tls_version;0:~15:tls_established;5:false!19:timestamp_tls_setup;0:~19:timestamp_tcp_setup;0:~15:timestamp_start;0:~13:timestamp_end;0:~14:source_address;0:~3:sni;0:~10:ip_address;0:~2:id;36:600f6313-534a-49bf-9bd5-b41562dbea78;4:alpn;0:~7:address;20:10:google.com;3:443#]}11:client_conn;410:10:proxy_mode;7:regular;11:cipher_list;0:]11:alpn_offers;0:]16:certificate_list;0:]3:tls;5:false!5:error;0:~8:sockname;7:0:;1:0#]5:state;1:3#11:tls_version;0:~14:tls_extensions;0:]15:tls_established;5:false!19:timestamp_tls_setup;0:~15:timestamp_start;18:1689604551.3295212^13:timestamp_end;0:~3:sni;0:~8:mitmcert;0:~2:id;36:77b8f50c-c5fe-45a6-9b9f-71b5b40256fe;11:cipher_name;0:~4:alpn;0:~7:address;7:0:;1:0#]}5:error;0:~2:id;36:ca39fbd9-7b44-4c13-859c-d2447c09d074;}1398:4:type;4:http;7:version;2:18#9:websocket;0:~8:response;0:~7:request;292:4:path;1:/,9:authority;0:,6:scheme;5:https,6:method;4:POST,4:port;3:443#4:host;10:google.com;13:timestamp_end;18:1689604562.9142652^15:timestamp_start;18:1689604562.9142652^8:trailers;0:~7:content;0:,7:headers;51:22:14:content-length,1:0,]21:4:Host,10:google.com,]]12:http_version;8:HTTP/1.1,}17:timestamp_created;17:1689604562.915139^7:comment;0:;8:metadata;0:}6:marked;0:;9:is_replay;0:~11:intercepted;5:false!11:server_conn;408:4:via2;0:~11:cipher_list;0:]11:cipher_name;0:~11:alpn_offers;0:]16:certificate_list;0:]3:tls;5:false!5:error;0:~5:state;1:0#3:via;0:~11:tls_version;0:~15:tls_established;5:false!19:timestamp_tls_setup;0:~19:timestamp_tcp_setup;0:~15:timestamp_start;0:~13:timestamp_end;0:~14:source_address;0:~3:sni;0:~10:ip_address;0:~2:id;36:ad48cb4c-ee61-4c22-bef5-c64dab14d678;4:alpn;0:~7:address;20:10:google.com;3:443#]}11:client_conn;410:10:proxy_mode;7:regular;11:cipher_list;0:]11:alpn_offers;0:]16:certificate_list;0:]3:tls;5:false!5:error;0:~8:sockname;7:0:;1:0#]5:state;1:3#11:tls_version;0:~14:tls_extensions;0:]15:tls_established;5:false!19:timestamp_tls_setup;0:~15:timestamp_start;18:1689604562.9141653^13:timestamp_end;0:~3:sni;0:~8:mitmcert;0:~2:id;36:e12f7a90-f8f5-4975-8292-638bb7b866ab;11:cipher_name;0:~4:alpn;0:~7:address;7:0:;1:0#]}5:error;0:~2:id;36:ab38d892-38fa-4b40-840e-466eb01629f1;}
